Organise classes into suitable packages
