+++ src/main/java/chacha/tasks/Task.java
@@ -1,3 +1,4 @@
+package chacha.tasks;
 public class Task {
     private String description;
     private boolean isDone;
